+++ src/app/app.module.ts
@@ -1,24 +1,35 @@
+import { CommonModule } from '@angular/common'; 
 import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { ErrorHandler, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
+import { MatTableModule } from '@angular/material/table';
+
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { DashboardComponent } from '../components/dashboard/dashboard.component';
+import { ErrorComponent } from '../components/error/error.component';
+import { PageNotFoundComponent } from '../components/page-not-found/page-not-found.component';
 import { GlobalErrorHandlerService } from '../services/global-error-handler.service';
 import { HttpErrorInterceptor } from '../interceptors/http-error-interceptor';
 
 @NgModule({
   declarations: [
     AppComponent,
-    DashboardComponent
+    DashboardComponent,
+    ErrorComponent,
+    PageNotFoundComponent
   ],
   imports: [
+    CommonModule,
     BrowserModule,
     AppRoutingModule,
     BrowserAnimationsModule,
-    HttpClientModule
+    HttpClientModule,
+
+    // Material
+    MatTableModule
   ],
   providers: [
     { provide: HTTP_INTERCEPTORS, useClass: HttpErrorInterceptor, multi: true },