+++ src/components/dashboard/dashboard.component.ts
@@ -1,15 +1,29 @@
 import { Component, OnInit } from '@angular/core';
 
+import { DataService } from '../../services/data.service'; 
+import { Superhero } from '../../interfaces/superhero.interface';
+
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html',
   styleUrls: ['./dashboard.component.scss']
 })
 export class DashboardComponent implements OnInit {
+  public columnsToDisplay: string[];
+  public superheroes: Superhero[] = [];
 
-  constructor() { }
+  constructor(
+    private readonly dataService: DataService
+  ) { }
 
   ngOnInit(): void {
+    // Keys from interface to show them as headers in the table
+    this.columnsToDisplay = ['nameLabel', 'genderLabel', 'citizenshipLabel', 'skillsLabel', 'occupationLabel', 'memberOfLabel', 'creatorLabel'];
+
+    // Get superheroes from the service
+    this.dataService.getSuperheroes().subscribe( (data: Superhero[]) => {
+      this.superheroes = data;
+    })
   }
 
 }