+++ src/interfaces/superhero.interface.ts
@@ -1,0 +1,9 @@
+export interface Superhero {
+  nameLabel: string;
+  genderLabel: string;
+  citizenshipLabel: string;
+  skillsLabel: string;
+  occupationLabel: string;
+  memberOfLabel: string;
+  creatorLabel: string;
+}